chore: finaliza icon utilizando o generate
