+++ packages/ocean-icons-react/src/index.ts
@@ -31,7 +31,6 @@ export { default as BookOpen } from './BookOpen';
 export { default as BookmarkAlt } from './BookmarkAlt';
 export { default as Bookmark } from './Bookmark';
 export { default as Briefcase } from './Briefcase';
-export { default as BtokenOutline } from './BtokenOutline';
 export { default as Cake } from './Cake';
 export { default as Calculator } from './Calculator';
 export { default as Calendar } from './Calendar';
@@ -279,6 +278,7 @@ export { default as BookOpenOutline } from './BookOpenOutline';
 export { default as BookmarkAltOutline } from './BookmarkAltOutline';
 export { default as BookmarkOutline } from './BookmarkOutline';
 export { default as BriefcaseOutline } from './BriefcaseOutline';
+export { default as BtokenOutlineOutline } from './BtokenOutlineOutline';
 export { default as CakeOutline } from './CakeOutline';
 export { default as CalculatorOutline } from './CalculatorOutline';
 export { default as CalendarOutline } from './CalendarOutline';
